Address Copilot review comments on sheet accessibility and sync
