+++ frontend/src/features/quotes/components/reviewLineItemSheetState.ts
@@ -1,7 +1,7 @@
 import type { QuoteEditDraft } from "@/features/quotes/hooks/useQuoteEdit";
 import type { LineItemDraftWithFlags } from "@/features/quotes/types/quote.types";
+import { syncDraftTotalWithLineItems } from "@/features/quotes/utils/lineItemDraftTotals";
 import { DOCUMENT_LINE_ITEMS_MAX_ITEMS } from "@/shared/lib/inputLimits";
-import { resolveLineItemSum } from "@/shared/lib/pricing";
 
 export type ReviewLineItemSheetState =
   | { mode: "add" }
@@ -65,43 +65,4 @@ export function applyLineItemSheetDelete(
     lineItems: nextLineItems,
     total: syncDraftTotalWithLineItems(draft, nextLineItems),
   };
-}
-
-function syncDraftTotalWithLineItems(
-  currentDraft: QuoteEditDraft,
-  nextLineItems: LineItemDraftWithFlags[],
-): number | null {
-  const currentDerivedSubtotal = resolveFullyPricedLineItemSum(currentDraft.lineItems);
-  if (currentDerivedSubtotal !== currentDraft.total) {
-    return currentDraft.total;
-  }
-
-  const nextDerivedSubtotal = resolveFullyPricedLineItemSum(nextLineItems);
-  if (nextDerivedSubtotal === null) {
-    return hasSubstantiveLineItems(nextLineItems) ? currentDraft.total : null;
-  }
-  return nextDerivedSubtotal;
-}
-
-function resolveFullyPricedLineItemSum(lineItems: LineItemDraftWithFlags[]): number | null {
-  const substantiveLineItems = lineItems.filter(hasLineItemContent);
-  if (substantiveLineItems.length === 0) {
-    return null;
-  }
-  if (substantiveLineItems.some((lineItem) => lineItem.price === null)) {
-    return null;
-  }
-  return resolveLineItemSum(substantiveLineItems.map((lineItem) => lineItem.price));
-}
-
-function hasSubstantiveLineItems(lineItems: LineItemDraftWithFlags[]): boolean {
-  return lineItems.some(hasLineItemContent);
-}
-
-function hasLineItemContent(lineItem: LineItemDraftWithFlags): boolean {
-  return (
-    lineItem.description.trim().length > 0
-    || (lineItem.details?.trim().length ?? 0) > 0
-    || lineItem.price !== null
-  );
 }
