+++ frontend/src/features/quotes/hooks/useQuoteEdit.ts
@@ -1,7 +1,8 @@
 import { useCallback, useState } from "react";
 
 import type { LineItemDraftWithFlags } from "@/features/quotes/types/quote.types";
-import { resolveLineItemSum, type DiscountType } from "@/shared/lib/pricing";
+import { syncDraftTotalWithLineItems } from "@/features/quotes/utils/lineItemDraftTotals";
+import { type DiscountType } from "@/shared/lib/pricing";
 
 const EDIT_STORAGE_KEY = "stima_quote_edit";
 
@@ -227,47 +228,4 @@ export function useQuoteEdit(): UseQuoteEditResult {
     removeLineItem,
     clearDraft,
   };
-}
-
-
-function syncDraftTotalWithLineItems(
-  currentDraft: QuoteEditDraft,
-  nextLineItems: LineItemDraftWithFlags[],
-): number | null {
-  const currentDerivedSubtotal = resolveFullyPricedLineItemSum(currentDraft.lineItems);
-  if (currentDerivedSubtotal !== currentDraft.total) {
-    return currentDraft.total;
-  }
-
-  const nextDerivedSubtotal = resolveFullyPricedLineItemSum(nextLineItems);
-  if (nextDerivedSubtotal === null) {
-    return hasSubstantiveLineItems(nextLineItems) ? currentDraft.total : null;
-  }
-  return nextDerivedSubtotal;
-}
-
-
-function resolveFullyPricedLineItemSum(lineItems: LineItemDraftWithFlags[]): number | null {
-  const substantiveLineItems = lineItems.filter(hasLineItemContent);
-  if (substantiveLineItems.length === 0) {
-    return null;
-  }
-  if (substantiveLineItems.some((lineItem) => lineItem.price === null)) {
-    return null;
-  }
-  return resolveLineItemSum(substantiveLineItems.map((lineItem) => lineItem.price));
-}
-
-
-function hasSubstantiveLineItems(lineItems: LineItemDraftWithFlags[]): boolean {
-  return lineItems.some(hasLineItemContent);
-}
-
-
-function hasLineItemContent(lineItem: LineItemDraftWithFlags): boolean {
-  return (
-    lineItem.description.trim().length > 0
-    || (lineItem.details?.trim().length ?? 0) > 0
-    || lineItem.price !== null
-  );
 }
